requirements.txt
torch
torchvision
flask
streamlit
pillow
numpy
pandas
